Make js component & fix summary
