+++ lang/en/page.php
@@ -18,4 +18,8 @@ return [
     'cart.quantity' => 'Quantity',
     'cart.empty' => 'Your cart is empty.',
     'cart.button' => 'Explore our products',
+
+    'checkout.pay' => 'Pay',
+    'checkout.transaction_fee' => 'Transaction Fee',
+    'checkout.shipping_fee' => 'Shipping Fee',
 ];
