+++ resources/views/checkout/component/__summary-card.blade.php
@@ -1,4 +1,4 @@
-@props(['subtotal', 'quantity'])
+@props(['subtotal', 'quantity', 'transaction', 'shipping', 'total'])
 
 <form action="{{ route('cart.checkout') }}" method="POST">
     @csrf
@@ -17,22 +17,22 @@
                 <p>Rp {{ $subtotal }}</p>
             </div>
             <div class="flex justify-between pb-2">
-                <p class="font-medium text-font_secondary">Transaction Fee</p>
-                <p>Rp </p>
+                <p class="font-medium text-font_secondary">{{ __('page.checkout.transaction_fee') }}</p>
+                <p>Rp {{ $transaction }}</p>
             </div>
             <div class="flex justify-between">
-                <p class="font-medium text-font_secondary">Shipping Fee</p>
-                <p>Rp </p>
+                <p class="font-medium text-font_secondary">{{ __('page.checkout.shipping_fee') }}</p>
+                <p>Rp {{ $shipping }}</p>
             </div>
         </div>
         <div class="p-4">
             <div class="flex justify-between font-semibold text-font_primary">
                 <p>Total</p>
-                <p>Rp </p>
+                <p>Rp {{ $total }}</p>
             </div>
             <button type="submit" class="mt-4 flex w-full justify-center items-center gap-4 text-white bg-button disabled:cursor-not-allowed disabled:bg-button/70 hover:bg-button/80 focus:ring-4 focus:outline-none focus:ring-button/15 font-medium rounded-lg text-sm px-5 py-2.5 text-center">
                 @include('component.__spinner')
-                Checkout
+                {{ __('page.checkout.pay') }}
             </button>
         </div>
     </div>
