+++ resources/views/component/js/__common-js.blade.php
