+++ resources/views/layout/layout.blade.php
@@ -5,7 +5,7 @@
     @include('component.navigation.navbar')
     @yield('content')
     @include('component.__footer')
-    @include('component.common-js')
+    @include('component.js.__common-js')
     @yield('extra-js')
 </body>
 </html>
